use mapPartitions rather than extra classes
The Filter operations (contains, intersects, ...) required only to iterate over a partition and apply the filter. This can easily be done with the RDD#mapPartitions transformation. So we don't have to maintain so many classes that differ in only the call of the predicate.

The classes for Joins and LiveIndexing are kept for now.