use a cartesian product approach for joins
previously, we added all n partitions of left with a dependency list of m partitions of right.
We now generator n*m partitions (i.e. all combinations). If partitioners are set, those n*m partitions
are pruned on generation, so that combinations that cannot contribute to the result are not generated,
and not processed later on