fix bottle neck for joins without index
When joining wihtout index, we use a simple nested loop approach in which we previously received the iterators from left and right inputs. However, it turned out that getting the iterator of the inner RDD again and again is very expensive. Thus, the inner iterator is fetched once and stored into a list which we then use for the inner loop.